Annotate a CDS by Label Transfer

Use this function to transfer cell labels from one single cell dataset to another. If a cds is provided for either reference or query, it is converted to a seurat object and the labels are transferred to the query by anchor finding. The assignments will be in the form of a new cds column with name predicted."column name from reference". This should be unique on the first application of the function to a query dataset. However, if running queries against more than 1 reference data set it is possible that you will unintentionally generate the same column name which would overwrite the first assignment column. The function checks for this and aborts with the recommendation to supply a unique id to the unique_id parameter.

Usage

bb_cds_anno(query_cds, ref, transfer_col, unique_id = NULL)

Arguments

query_cds

The single cell data set you wish to annotate. Must be a CDS.

ref

The reference single cell data set. May be either a CDS or Seurat object.

transfer_col

The column from the reference data set that provides the labels.

unique_id

A unique identifier to add to the column with the transferred labels. Default is NULL but it is recommended to provide an informative label when annotating against more than one reference. Default is NULL.

Value

a CDS with two new cell metadata columns

See also