Make an Ape Transcriptome Object

This function takes a specific ensembl transcript identifier, such as ENST00000348343.11, and gets the cDNA sequence from the corresponding transcriptomic reference. This is returned as an Ape object with the UTR's and the CDS annotated as features.

Usage

bb_make_ape_transcript(query, transcriptome = c("hg38", "GRCz11"))

Arguments

query

A specific ensembl transcript identifier.

transcriptome

Genome/transcriptome reference to use, Default: c("hg38", "GRCz11")

Value

OUTPUT_DESCRIPTION

Details

DETAILS

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 }
} # }