Create a Loupe File from a Cell Data Set

Converts a cell data set into a loupe file.

Usage

bb_loupeR(cds, output_dir = ".", output_file = "loupe")

Arguments

cds

Input cds. Only works on cell data set objects. For seurat objects, use built in loupeR functions

output_dir

Output directory, Default: '.'

output_file

Name of the loupe file. .cloupe will be appended for compatibility, Default: 'loupe'

Value

Nothing

See also